requirements.txt
langchain
langgraph 
langchainhub 
langchain-community
python-dotenv 
langchain-openai
git+https://github.com/PyGithub/PyGithub.git
gitpython
streamlit
fastapi
uvicorn